CI: Add News Check
We are using `towncrier` to generate the changelog for this project, and
`towncrier` reads changes out of the `news/` directory.  This patch ensures
that new PRs that change the functionality of the SDK either come with a file
in the `news/` directory that explains their changes or is explicitly marked
with a `skip news` tag.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>