#2 Add GitHub Actions to Python SDK
* CI: Update tests so they lint properly

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>

* CI: Add DCO Check for PRs

This project requires contributors to verify that they are releasing
their changes under the Apache 2.0 license that we use, by signing off
on a DCO.  This patch adds a CI check that ensures each commit in a PR
has a `Signed-off-by` trailer indicating that the contributor has agreed
to our DCO.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>

* CI: Add News Check

We are using `towncrier` to generate the changelog for this project, and
`towncrier` reads changes out of the `news/` directory.  This patch ensures
that new PRs that change the functionality of the SDK either come with a file
in the `news/` directory that explains their changes or is explicitly marked
with a `skip news` tag.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>

* CI: Add action to build and test the library

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>

* CI: Add an environment variable to turn off pkg-config checks

We want to be able to generate sdists from our `setup.py` without requiring
BlazingMQ to be on the system (as it is not, for instance, on a GitHub Action
container).  This patch introduces an environment variable for this purpose.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>

* CI: Build wheels

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>

---------

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>