docs: Explicitly document underlying `BMQ_BROKER_URI` behavior (#4)
`libbmq` supports a `BMQ_BROKER_URI` environment variable, which when set,
[overrides any broker URI passed in through the `SessionOptions`][envvar].  As
the Python SDK is a set of bindings on top of `libbmq`, we inherit this
behavior as well.  More specifically, setting `BMQ_BROKER_URI` overrides
whatever value is passed in with the `broker` argument to the
`blazingmq.Session` object.  Our integration tests use the environment variable
in this way, so it is important to document it correctly.

This patch documents this behavior, which is stable and supported in
`libbmq`, is exposed through the Python SDK, and is used by the Python
SDK’s tests.

[envvar]: https://bloomberg.github.io/blazingmq/docs/apidocs/cpp_apidocs/group__bmqa__session.html

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>