ci: Turn off docs publishing job (#3)
Because we host our Python API documentation on the main BlazingMQ repository,
we are not able to use a GitHub action as simple as this to generate and update
the documentation.  This patch removes this action.  For the moment, we will
update the documentation manually on release, but we may bring this back in a
smarter way, or move the SDK documentation into this repository in the future.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>