Add `Timeouts` class for setting each operation’s timeout
Currently, the Python SDK does not allow users to specify different timeouts
for each session operation, instead accepting a single floating-point value
that sets open, configure, and close queue timeouts.  This is in contrast to
the C++ `libbmq` SDK, which allows users to configure timeouts for each
operation separately using the `bmqt_sessionoptions` component.

In order to preserve the behavior of existing clients, while enabling feature
parity between the Python SDK and the C++ SDK with regard to selecting
timeouts, this patch introduces a new value-semantic class, `Timeouts`, which
allows users to independently set the timeouts for each of the five session
operations whose timeouts are configureable: session connection, session
disconnection, queue opening, queue configuration, and queue closing.  It is
intended that clients may either continue to provide a `float` value with the
same semantics when creating a `Session` object, or may instead provide an
instance of `Timeouts` for the same parameter value to configure each timeout
separately.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>