Expose more session options through `_session.py`
This patch completes the trio of patches that expose libbmq’s new
`SessionOptions` to the users of the Python SDK through the latter’s three
layers, by allowing users to build Python `Session`s that specify all the new
session options.

The changes in this patch should nonetheless maintain backwards compatibility
with all existing clients of the Python SDK.  All new session options that the
`Session` constructor accepts are defaulted to `None` and as such are not set
by pybmq, maintaining the existing behavior for callers who do not set any of
these new arguments.  New session options are also appended to the list of
arguments, so clients who build a `Session` using positional arguments rather
than named arguments will continue to function correctly.  Finally, by
overloading the `timeout` argument to taking either a simple float or an
instance of the new value-semantic class `Timeouts`, users who currently set a
timeout for the `Session` by passing in a single timeout will continue to see
the same behavior, but users passing in a new `Timeouts` value have access to
the more fine-grained timeout settings that libbmq’s `SessionOptions` provides.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>