Add `SessionOptions` class to Python layer
While users can continue to construct the `Session` directly using its
constructor, adding the additional session options arguments to the constructor
has made it somewhat unwieldly to call, especially if the user only wants to
configure one of the session options.  This patch provides an alternative,
simpler way to construct a `Session` instance, using a new value-semantic
builder class named `SessionOptions` that holds each of the new session
options and a class method named `with_options`, which takes an instance of
`SessionOptions` and returns a new `Session`.  `with_options` is the
recommended way to construct a new session if you plan to configure the session
options.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>