Update Python `MockSession`
The `MockSession` class in the Python SDK is used by our test suite to ensure
that session options are properly propagated through our Cython layer so they
can be passed to libbmq.  However, because have introduced new session options
into the Python SDK, the `MockSession` class needs to be updated so it knows
about these additional options.  This patch extends `MockSession` with the new
session options.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>