CI: Install broker on MacOS builds
This patch brings the `build-macos-universal.sh` script to parity with the
`build-manylinux.sh` script in installing the BlazingMQ broker to `DIR_INSTALL`
after building it.  This is necessary to run integration tests on an OSX GitHub
Actions runner, and better automates the process for running these tests
manually, bringing the process up to line with the process on Linux machines.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>