typing: Use `Mapping` for proper variance
The `Dict[K,V]` type is invariant in both `K` and `V`, so if `V1` is a subtype
of `V2`, it isn’t the case that `Dict[K,V1]` is not a subtype of `Dict[K,V2]`.
This patch changes type annotations from the concrete type `Dict` to the
abstract base class `Mapping`, which is covariant in `V`, fixing some correct
uses in our package that are mistakenly marked as incorrect by type checkers.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>