typing: Add hint to typechecker in `get_script_name`
This function looks up the name of the Python script, so as to provide it to
the `NativeSession` as an identifier.  We guard the use of this name with a
`getattr` conditional, but the typechecker cannot determine that this forces it
to be non-`None`, so this patch adds an additional `assert` which never fires,
but which the typechecker can understand to silence an error.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>