typing: Fix incorrect type annotations in `on_message`
The type annotation for `on_message`’s `messages` argument has been incorrect,
and hasn’t matched the values that are being passed to it in tests.
`on_message` performs a conversion from the BlazingMQ wire protocol identifiers
for message property types, which are represented as `int`s, to the Python
SDK’s message property type enumeration, which are Python objects.  However,
the type annotation incorrectly documented that this conversion had already
taken place.  This patch updates the type annotation and makes the code that
does the conversion clearer.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>