CI: Update GitHub Actions to use Node.js 20
Our CI has been admonishing us to use more recent versions of certain GitHub
Actions, like `actions/checkout`, which are built on Node.js version 20 instead
of the older Node.js version 16.  This transition is documented
[here][actions-transition].

This patch updates the actions we can to ones that are built on Node.js 20:

  - `actions/checkout@v3`          → `actions/checkout@v4`
  - `actions/upload-artifact@v3`   → `actions/upload-artifact@v4`
  - `actions/download-artifact@v3` → `actions/download-artifact@v4`
  - `actions/setup-python@v4`      → `actions/setup-python@v5`
  - `docker/setup-qemu-action@v2`  → `actions/setup-qemu-action@v3`

This patch also uses this opportunity to update our dependency on
`pypa/cibuildwheel` from `v2.15.0` to the latest version, `v2.16.2`.

[actions-transition]: https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>