CI: Explicitly separate `dist` archives and remerge them
Upgrading to `actions/upload-artifact@v4` to use version 20 of NodeJS caused
our entire CI wheel-building matrix to fail.  It seems that `v4` of this action
introduced a breaking change when uploading to the an artifact with the same
name several times during the same workflow: in `v3`, this would silently
incorporate the new files into the same artifact; in `v4`, however, artifacts
are immutable, and the job simply fails.  It is possible [to migrate
this behavior][v4-migration] by uniquely renaming each uploaded archive, and
reuploading a single artifact with the new `actions/upload-artifact/merge@v4`
action.  This patch renames all artifacts created when building sdists, Linux
wheels, and MacOS wheels, and then introduces a new job to merge each of these
separate artifacts into a single `dist` archive.

[v4-migration]: https://github.com/actions/upload-artifact/blob/main/docs/MIGRATION.md

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>