CI: Allow PyPI publishing to be done manually
It is useful to be able to dispatch a PyPI upload manually, without creating a
new GitHub release, in the case of transient GitHub Actions failures that are
not due to a problem in the release itself.  This patch enables the
`workflow_dispatch` trigger for our `wheels`, and explicitly allows
`upload_pypi` to run on `workflow_dispatch`, which allows us to manually and
intentionally call this workflow without making a new release.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>