CI: Allow PyPI publishing to be done manually
We have occassionally gotten transient GitHub Action failures that seem
to be due to GitHub’s infrastructure, rather than our own workflows.  In
the case of creating a new GitHub release, this may result in a failure
to upload the build wheels to PyPI.  For that reason, it is useful to be
able to dispatch a PyPI upload manually, without creating a new GitHub
release.  This patch enables the `workflow_dispatch` trigger for our
`wheels`, and explicitly allows `upload_pypi` to run on
`workflow_dispatch`, which allows us to manually and intentionally call
this workflow without making a new release.

We want this to be run intentionally, though, so it’s important to note
that to trigger a `workflow_dispatch` event, the workflow has to be on
`main`.  This means we cannot publish from a branch or PR that has not
yet been merged.  Because the primary use case of this is to try
publishing a PyPI upload again after a transient failure due to GitHub’s
infrastructure, this restriction is useful.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>