[WIP] CI: Automatically publish docs to GitHub Pages
Just as we upload to PyPI on new releases (as well as on manual dispatch of the
wheel-building workflow), we also want to build and upload documentation on new
releases.  We are moving to hosting the documentation locally on our own GitHub
pages, so we can do this automatically, rather than invoke a PR on the main
`blazingmq` repository.

This patch is a work-in-progress, and should not be merged yet.  We want to
ensure that the documentation is built correctly before we attempt to publish
it, so we run this on every build of the wheels, but also only publish to
GitHub Pages as a dry run.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>