Document `QueueOptions` behavior with 0 `max_unconfirmed`
Our examples in `examples/consumer.py` and `examples/consumer2.py`
encourage users to start a session with `queue_options=QueueOptions(0,
0, 0)`.  This expression is relatively terse, but it means:

    queue_options=QueueOptions(
        max_unconfirmed_messages=0,
        max_unconfirmed_bytes=0,
        consumer_priority=0
    )

Our documentation for `max_unconfirmed_messages` and
`max_unconfirmed_bytes` do not mention what happens when they are set
to 0; the clearest reading is that they do not allow *any* unconfirmed
messages, which means they should not be able to accept any messages
at all!  However, the examples work fine.

A quick investigation reveals a few things.  First, neither
`DEFAULT_MAX_UNCONFIRMED_MESSAGES` nor `DEFAULT_MAX_UNCONFIRMED_BYTES`
are 0.  Set in ‘bmqt_queueoptions` in the BlazingMQ C++ SDK, these two
constants have the values 1000 and 33554432, respectively.  So, the
above expression yields a different `QueueOptions` value than the
following:

    queue_options=QueueOptions(
        max_unconfirmed_messages=None,
        max_unconfirmed_bytes=None,
        consumer_priority=0
    )

Second, a value of 0 for either one of these options does not mean
that no unconfirmed messages are allowed.  Instead, the documentation
for the corresponding `bmqt_queueoptions` component indicates that 0
for either of these indicates that no limit on the number of
unconfirmed messages or the maximum size in bytes of unconfirmed
messages is imposed.  Because we propagate this value from the Python
`QueueOptions` to the C++ `bmqt_queueoptions` component relatively
directly, the behavior on 0 is the same for Python.

This patch explicitly documents the behavior of setting
`QueueOptions`’s `max_unconfirmed_messages` or `max_unconfirmed_bytes`
to 0.  Making this change illuminates the behavior of the relatively
opaque call in the example consumers we provide.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>