tests: Extend producer example test with additional checks
This patch extends the `test_examples.py` integration test for the
simple producer example with two additional checks inspired by the new
producer-with-correlation example test.  First, we make sure that the
producer process completes successfully.  Second, we verify that it only
sent one message to the consumer while it was alive.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>