docs: Document OSX test crash behavior
Depending how your system is configured, running the
`test_deadlock_detection.py` integration test on OSX may result in a
dialog window warning you about a crashed or aborted Python process.
This crash is intentional, as the test introduces a deadlock and tests
whether it can be detected.  This patch documents this behavior in both
the build documentation and the test itself.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>