examples: Add example of producer with post/ack correlation
The BlazingMQ Python SDK and the BlazingMQ C++ SDKs use two different
mechanisms for correlating a message that a producer posts with the ack
or nack that the broker sends back.  The BlazingMQ C++ SDK features a
“correlation id”, which is a value that a producer can provide along
with a posted message and that the broker will return with that
message’s ack or nack.  This concept is quite conspicuous in the C++ API
and is featured in the advanced producer tutorial that is provided along
with the SDK.  In the Python SDK, on the other hand, there is no
corresponding concept; we do not expose the C++ correlation id to
clients using the Python SDK.  In fact, internally, we use the
correlation ID to store the pointer to the `on_ack` callback that the
user provides with each posted message.

Several users have asked how to correlate a message’s `post` call with
its ack or nack.  The expected idiom to do this in Python is to pass a
different closure as the `on_ack` callback for each message you want to
correlate.  These closures can store arbitrary data that can be used in
the `on_ack` callback, allowing producers to store an identifier by
which they can correlate posts and acks/nacks.

This patch includes a new example producer, `producer_with_correlation`
that demonstrates this idiom, posting ten messages and providing each
one with a distinct `on_ack` closure, constructed using the light-weight
`functools.partial`.  The description for this test is included in the
examples documentation, and to make sure that this producer works as
expected, we also add it to the `test_examples.py` integration test.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>